Refactor row ID reassignment planning
